agrego sensor de 0-180 para las aguas

+++ Truma_EBL_Bios_V100/Truma_EBL_Bios_V100/Core/Inc/FreshWater_1.h
@@ -10,7 +10,7 @@ private:
     unsigned long long expirationTime;
 		unsigned long long powerTime;
     int state;
-		int levelFW;
+		float levelFW;
 		int counterFW;
 public:
     FreshWater1(short _idTopic);
