+++ Truma_EBL_Bios_V100/Truma_EBL_Bios_V100/Core/Inc/GrayWater_1.h
@@ -10,7 +10,7 @@ private:
     unsigned long long expirationTime;
 		unsigned long long powerTime;
     int state;
-		int levelGW;
+		float levelGW;
 		int counterGW;
 public:
     GrayWater1(short _idTopic);
